backend: refactor: move pg.js, setupScraping.js, and setupAirtableBackup.js to db/ directory
